Remove GetEstimatedMaxItemCount()
This patch removes the method Index::GetEsitmatedMaxItemCount().
The original purpose was that the index provides an upper limit on the
number of stored items.

However, the idea went bad in tons of situations:
1) Items do not always have the same size. The size is more or less
static for chunk mappings, up block recipes differ significantly.
2) Often, the storage usage of a index is not static. The LSM-trees
   can shrink even if no data is deleted, due to compactions. There
   is no real way to provide an upper limit.
3) Most real implementations just forwarded a fixed configurable number
   that was often used highly conservative.
4) It is often not possible to provide a safe estimate for the storage
   overhead.

This is the first breaking patch for the dedupv1 2.0 development
line.

The better approach that dedupv1 should pursue in the future
development is to handle the file system out ot space messages.
This should even today never provide wrong results in out of space
situations.