Remove the parallel filter chain execution.
In a dedupv1 1.0 the filter chains of all chunks are executed in
parallel. This is a) no longer necessary in a dedupv1 2.0 setting, b)
hurts performance (20 MB/s jump) and c) is more complex.

So I would like to remove it.

This is a change that probably will hurt the performance in dedupv1 1.0
configurations.