Remove all persistent tokyo cabinet indexes.
The performance of these indexes are not been sufficent.
Sqlite and LevelDB provide a significantly better performance.
Therefore this patch removes these index implementations.

The in-memory tc index implementations are keep as they are still
used for in-memory indexes, e.g. the block index auxiliary index.
Right now there is no good replacement for them.